Fix UserCodeDeployment to work with inner classes
The problem with only core side user code deployment.
Tests are added to the client as well.

Also cleaned up UserCodeDeploymentAbstractTest which
were used for the compatibility tests on enterprise repo.
Compatibility tests are removed since they are not needed
anymore after switching to 4.0.

Following two tests are newly added to `UserCodeDeploymentBasicTest` which are related to this fix
testInnerClassFetchedFirst_thenMainClassFetchedFromRemote
testMainClassFetchedFirst_thenInnerlassFetchedFromRemote

Others are moved from `UserCodeDeploymentAbstractTest` when refactoring.

fixes https://github.com/hazelcast/hazelcast/issues/17044
