+++ hazelcast/src/main/java/com/hazelcast/internal/usercodedeployment/impl/ClassLocator.java
@@ -162,13 +162,13 @@ public final class ClassLocator {
                     classSource = doPrivileged(
                             (PrivilegedAction<ClassSource>) () -> new ClassSource(parent, this, Collections.emptyMap()));
                 }
-                ClassData classData = fetchBytecodeFromRemote(name);
+                ClassData classData = fetchBytecodeFromRemote(mainClassName);
                 if (classData == null) {
                     throw new ClassNotFoundException("Failed to load class " + name + " from other members");
                 }
 
                 Map<String, byte[]> innerClassDefinitions = classData.getInnerClassDefinitions();
-                classSource.define(name, classData.getMainClassDefinition());
+                classSource.define(mainClassName, classData.getMainClassDefinition());
                 for (Map.Entry<String, byte[]> entry : innerClassDefinitions.entrySet()) {
                     classSource.define(entry.getKey(), entry.getValue());
                 }
