+++ hazelcast/src/test/java/com/hazelcast/client/usercodedeployment/ClientUserCodeDeploymentTest.java
@@ -27,7 +27,8 @@ import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.internal.util.FilteringClassLoader;
 import com.hazelcast.map.EntryProcessor;
 import com.hazelcast.map.IMap;
-import com.hazelcast.test.HazelcastSerialParametersRunnerFactory;
+import com.hazelcast.test.HazelcastParallelParametersRunnerFactory;
+import com.hazelcast.test.annotation.ParallelJVMTest;
 import com.hazelcast.test.annotation.QuickTest;
 import org.junit.After;
 import org.junit.Test;
@@ -37,6 +38,7 @@ import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
 import org.junit.runners.Parameterized.UseParametersRunnerFactory;
 import usercodedeployment.CapitalizingFirstNameExtractor;
+import usercodedeployment.DomainClassWithInnerClass;
 import usercodedeployment.EntryProcessorWithAnonymousAndInner;
 import usercodedeployment.IncrementingEntryProcessor;
 import usercodedeployment.Person;
@@ -53,8 +55,8 @@ import static java.util.Collections.singletonList;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(Parameterized.class)
-@UseParametersRunnerFactory(HazelcastSerialParametersRunnerFactory.class)
-@Category({QuickTest.class})
+@UseParametersRunnerFactory(HazelcastParallelParametersRunnerFactory.class)
+@Category({QuickTest.class, ParallelJVMTest.class})
 public class ClientUserCodeDeploymentTest extends ClientTestSupport {
 
     private TestHazelcastFactory factory = new TestHazelcastFactory();
@@ -265,6 +267,30 @@ public class ClientUserCodeDeploymentTest extends ClientTestSupport {
 
          */
         clientUserCodeDeploymentConfig.addJar("ChildParent.jar");
+        clientConfig.setUserCodeDeploymentConfig(clientUserCodeDeploymentConfig.setEnabled(true));
+
+        factory.newHazelcastInstance(createNodeConfig());
+        factory.newHazelcastClient(clientConfig);
+    }
+
+    @Test
+    public void testWithMainAndInnerClassesWorksIndependentOfOrder_withInnerFirst() {
+        ClientConfig clientConfig = new ClientConfig();
+        ClientUserCodeDeploymentConfig clientUserCodeDeploymentConfig = new ClientUserCodeDeploymentConfig();
+        clientUserCodeDeploymentConfig.addClass(DomainClassWithInnerClass.InnerClass.class);
+        clientUserCodeDeploymentConfig.addClass(DomainClassWithInnerClass.class);
+        clientConfig.setUserCodeDeploymentConfig(clientUserCodeDeploymentConfig.setEnabled(true));
+
+        factory.newHazelcastInstance(createNodeConfig());
+        factory.newHazelcastClient(clientConfig);
+    }
+
+    @Test
+    public void testWithMainAndInnerClassesWorksIndependentOfOrder_withMainFirst() {
+        ClientConfig clientConfig = new ClientConfig();
+        ClientUserCodeDeploymentConfig clientUserCodeDeploymentConfig = new ClientUserCodeDeploymentConfig();
+        clientUserCodeDeploymentConfig.addClass(DomainClassWithInnerClass.class);
+        clientUserCodeDeploymentConfig.addClass(DomainClassWithInnerClass.InnerClass.class);
         clientConfig.setUserCodeDeploymentConfig(clientUserCodeDeploymentConfig.setEnabled(true));
 
         factory.newHazelcastInstance(createNodeConfig());
